Update namespace and folder
