+++ composer.json
@@ -19,7 +19,7 @@
     },
     "autoload": {
         "psr-4": {
-            "OAPFW\\": "includes"
+            "Automattic\\WooCommerce\\OAPF\\": "src"
         }
     },
     "scripts": {
