+++ phpcs.xml.dist
@@ -37,7 +37,7 @@
 	</rule>
 
 	<rule ref="WordPress.Files.FileName">
-		<exclude-pattern>includes/*</exclude-pattern>
+		<exclude-pattern>src/*</exclude-pattern>
 	</rule>
 
 	<rule ref="WordPress.WP.Capabilities">
